move all tests from syft/new_tests to syft/tests add regex to skip fixture.py in pre-commit hook
