--- packages/syft/new_tests/__init__.py
@@ -1,2 +1,0 @@
-# relative
-from . import conftest  # noqa:F401
